Спринт 12. 4. История поиска сохраняется в SharedPreferences, работа с историей поиска вынесена в отдельный класс.
И еще пока не совсем верно работает сохранение экрана поиска при повороте и т.п.
Но это пока не требуется по заданию.

Еще хочу немного переделать установку visible|gone для элементов окна поиска. Все работает, но хочется элегантнее сделать.
