+++ app/src/main/java/com/example/playlistmaker/App.kt
@@ -1,8 +1,11 @@
 package com.example.playlistmaker
 
 import android.app.Application
+import android.content.SharedPreferences
 import android.content.res.Configuration
+import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
+
 import androidx.appcompat.widget.SwitchCompat
 
 const val PLAYLISTMAKER_PREFERENCES = "playlist_preferences" // ключ для SharedPreferences
@@ -18,12 +21,16 @@ class App : Application() {
             field = value
         }
 
+    // Переменная для работы с SharedPreferences (хранилище настроек)
+   // private lateinit var sharedPrefs: SharedPreferences
+
     override fun onCreate() {
 
         // Получаем тему приложения, выбранную пользователем, из SharedPreferences, а если
         // ничего туда не успели сохранить, то применим текущую тему приложения
-        val sharedPrefs = getSharedPreferences(PLAYLISTMAKER_PREFERENCES, MODE_PRIVATE)
+        var sharedPrefs = getSharedPreferences(PLAYLISTMAKER_PREFERENCES, MODE_PRIVATE)
         // MODE_PRIVATE - чтобы данные были досутпны только нашему приложению
+
         var stringFromSharedPrefs = sharedPrefs.getString(THEME_SWITCH_KEY, "")
         when (stringFromSharedPrefs) {
             "false" -> darkTheme = false
