+++ app/src/main/java/com/example/playlistmaker/SearchActivity.kt
@@ -28,82 +28,7 @@ import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
-var trackListSearchHistory: ArrayList<Track> = arrayListOf()
-
-/*(Track( // 1 элемент
-    trackId = 1,
-    trackName = "Smells Like Teen Spirit",
-    artistName = "Nirvana",
-    trackTimeMillis = 3000000,
-    artworkUrl100 = "https://is5-ssl.mzstatic.com/image/thumb/Music115/v4/7b/58/c2/7b58c21a-2b51-2bb2-e59a-9bb9b96ad8c3/00602567924166.rgb.jpg/100x100bb.jpg"
-),
-Track( // 2 элемент
-    trackId = 2,
-    trackName = "Billie Jean",
-    artistName = "Michael Jackson",
-    trackTimeMillis = 261000,
-    artworkUrl100 = "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/3d/9d/38/3d9d3811-71f0-3a0e-1ada-3004e56ff852/827969428726.jpg/100x100bb.jpg"
-),
-Track( // 3 элемент
-    trackId = 3,
-    trackName = "Stayin' Alive",
-    artistName = "Bee Gees",
-    trackTimeMillis = 246000,
-    artworkUrl100 = "https://is4-ssl.mzstatic.com/image/thumb/Music115/v4/1f/80/1f/1f801fc1-8c0f-ea3e-d3e5-387c6619619e/16UMGIM86640.rgb.jpg/100x100bb.jpg"
-),
-Track( // 4 элемент
-    trackId = 4,
-    trackName = "Whole Lotta Love",
-    artistName = "Led Zeppelin",
-    trackTimeMillis = 319800,
-    artworkUrl100 = "https://is2-ssl.mzstatic.com/image/thumb/Music62/v4/7e/17/e3/7e17e33f-2efa-2a36-e916-7f808576cf6b/mzm.fyigqcbs.jpg/100x100bb.jpg"
-),
-Track( // 5 элемент
-    trackId = 5,
-    trackName = "Sweet Child O'Mine",
-    artistName = "Guns N' Roses",
-    trackTimeMillis = 301800,
-    artworkUrl100 = "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/a0/4d/c4/a04dc484-03cc-02aa-fa82-5334fcb4bc16/18UMGIM24878.rgb.jpg/100x100bb.jpg"
-),
-Track( // 6 элемент
-    trackId = 6,
-    trackName = "Smells Like Teen Spirit 2",
-    artistName = "Nirvana",
-    trackTimeMillis = 3000000,
-    artworkUrl100 = "https://is5-ssl.mzstatic.com/image/thumb/Music115/v4/7b/58/c2/7b58c21a-2b51-2bb2-e59a-9bb9b96ad8c3/00602567924166.rgb.jpg/100x100bb.jpg"
-),
-Track( // 7 элемент
-    trackId = 7,
-    trackName = "Billie Jean 2",
-    artistName = "Michael Jackson",
-    trackTimeMillis = 261000,
-    artworkUrl100 = "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/3d/9d/38/3d9d3811-71f0-3a0e-1ada-3004e56ff852/827969428726.jpg/100x100bb.jpg"
-),
-Track( // 8 элемент
-    trackId = 8,
-    trackName = "Stayin' Alive 2",
-    artistName = "Bee Gees",
-    trackTimeMillis = 246000,
-    artworkUrl100 = "https://is4-ssl.mzstatic.com/image/thumb/Music115/v4/1f/80/1f/1f801fc1-8c0f-ea3e-d3e5-387c6619619e/16UMGIM86640.rgb.jpg/100x100bb.jpg"
-),
-Track( // 9 элемент
-    trackId = 9,
-    trackName = "Whole Lotta Love 2",
-    artistName = "Led Zeppelin",
-    trackTimeMillis = 319800,
-    artworkUrl100 = "https://is2-ssl.mzstatic.com/image/thumb/Music62/v4/7e/17/e3/7e17e33f-2efa-2a36-e916-7f808576cf6b/mzm.fyigqcbs.jpg/100x100bb.jpg"
-),
-Track( // 10 элемент
-    trackId = 10,
-    trackName = "Sweet Child O'Mine 2",
-    artistName = "Guns N' Roses",
-    trackTimeMillis = 301800,
-    artworkUrl100 = "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/a0/4d/c4/a04dc484-03cc-02aa-fa82-5334fcb4bc16/18UMGIM24878.rgb.jpg/100x100bb.jpg"
-))*/
-// создаем адаптер для Track для истории поиска
-var trackAdapterSearchHistory = TrackAdapter()
-
-class SearchActivity : AppCompatActivity() {
+class SearchActivity : AppCompatActivity(), TrackViewHolder.OnItemClickListener {
     private var editString: String = ""
     private var trackList: ArrayList<Track> = arrayListOf()
 
@@ -124,33 +49,32 @@ class SearchActivity : AppCompatActivity() {
     // создаем адаптер для Track
     private val trackAdapter = TrackAdapter()
 
+    lateinit var searchHistory: SearchHistory //= SearchHistory(getSharedPreferences(PLAYLISTMAKER_PREFERENCES, MODE_PRIVATE))         // MODE_PRIVATE - чтобы данные были доступны только нашему приложению
+
     @SuppressLint("MissingInflatedId")
     override fun onCreate(savedInstanceState: Bundle?) {
 
         super.onCreate(savedInstanceState)
         setContentView(layout.activity_search)
 
+        searchHistory = SearchHistory(
+            getSharedPreferences(
+                PLAYLISTMAKER_PREFERENCES,
+                MODE_PRIVATE
+            )
+        )         // MODE_PRIVATE - чтобы данные были доступны только нашему приложению
+        searchHistory.trackAdapterSearchHistory.onItemClickListener = this
+
+        trackAdapter.onItemClickListener = this
+
         // LinearLayout истории поиска
         val linearLayoutSearchHistory = findViewById<LinearLayout>(id.searchHistoryLinearLayout)
 
         // Получаем историю поиска из SharedPreferences, а если
         // ничего туда не успели сохранить, то список пустой и не отображается
-        val sharedPrefs = getSharedPreferences(PLAYLISTMAKER_PREFERENCES, MODE_PRIVATE)
-        // MODE_PRIVATE - чтобы данные были доступны только нашему приложению
-        val json = sharedPrefs.getString(SEARCH_HISTORY_KEY, "")
+        if (searchHistory.trackListSearchHistory.isEmpty())
+            linearLayoutSearchHistory.visibility = View.GONE
 
-        when (json) {
-            "" -> {
-                trackListSearchHistory.clear()
-
-                linearLayoutSearchHistory.visibility = View.GONE
-            }
-
-            else -> {
-                val trackListType = object : TypeToken<ArrayList<Track>>() {}.type
-                trackListSearchHistory = Gson().fromJson(json, trackListType)
-            }
-        }
 
         // кнопка "назад"
         val buttonSearchBack = findViewById<ImageView>(id.button_search_back)
@@ -171,11 +95,11 @@ class SearchActivity : AppCompatActivity() {
         // для истории поиска
         val rvItemsSearchHistory: RecyclerView = findViewById(R.id.rvSearchHistoryItems)
         rvItemsSearchHistory.apply {
-            adapter = trackAdapterSearchHistory
+            adapter = searchHistory.trackAdapterSearchHistory
             layoutManager =
                 LinearLayoutManager(this@SearchActivity, LinearLayoutManager.VERTICAL, false)
         }
-        trackAdapterSearchHistory.items = trackListSearchHistory
+        // trackAdapterSearchHistory.items = trackListSearchHistory
 
         // кнопка "Обновить"
         val buttonRefresh = findViewById<Button>(id.buttonRefresh)
@@ -197,8 +121,6 @@ class SearchActivity : AppCompatActivity() {
             val placeholderLayout: LinearLayout = findViewById(id.placeholderLinearLayout)
             placeholderLayout.visibility = View.GONE
 
-       //     val linearLayoutSearchHistory = findViewById<LinearLayout>(id.searchHistoryLinearLayout)
-       //     linearLayoutSearchHistory.visibility = View.VISIBLE
         }
 
         val simpleTextWatcher = object : TextWatcher {
@@ -212,13 +134,11 @@ class SearchActivity : AppCompatActivity() {
 
                 val linearLayoutSearchHistory =
                     findViewById<LinearLayout>(id.searchHistoryLinearLayout)
-                //linearLayoutSearchHistory.visibility = View.GONE
                 linearLayoutSearchHistory.visibility =
-                    if (inputEditText.hasFocus() && (s?.isEmpty()  == true) && !trackListSearchHistory.isEmpty())
+                    if (inputEditText.hasFocus() && (s?.isEmpty() == true) && !searchHistory.trackListSearchHistory.isEmpty())
                         View.VISIBLE
                     else View.GONE
 
-                //(s?.isEmpty()  == true)
                 trackList.clear()
                 trackAdapter.items = trackList
                 trackAdapter.notifyDataSetChanged()
@@ -283,10 +203,9 @@ class SearchActivity : AppCompatActivity() {
         // отображаем LinearLayout истории поиска, если фокус находится в inputEditText и inputEditText пуст
         inputEditText.setOnFocusChangeListener { view, hasFocus ->
             linearLayoutSearchHistory.visibility =
-                if (hasFocus && inputEditText.text.isEmpty() && !trackListSearchHistory.isEmpty())
+                if (hasFocus && inputEditText.text.isEmpty() && !searchHistory.trackListSearchHistory.isEmpty())
                     View.VISIBLE
                 else View.GONE
-            //   Log.w("hasFocus", hasFocus.toString())
         }
 
         inputEditText.requestFocus()
@@ -294,23 +213,12 @@ class SearchActivity : AppCompatActivity() {
         // кнопка "Очистить историю"
         val buttonCleanSearchHistory = findViewById<Button>(id.buttonCleanSearchHistory)
         buttonCleanSearchHistory.setOnClickListener {
-            trackListSearchHistory.clear()
-            trackAdapterSearchHistory.items = trackListSearchHistory
-            trackAdapterSearchHistory.notifyDataSetChanged()
+
+            searchHistory.clean()
 
             linearLayoutSearchHistory.visibility = View.GONE
 
-            // Записываем историю поиска в SharedPreferences
-            val sharedPrefs = it.context.getSharedPreferences(
-                PLAYLISTMAKER_PREFERENCES,
-                AppCompatActivity.MODE_PRIVATE
-            )
-
-            val json = Gson().toJson(trackListSearchHistory)
-            sharedPrefs.edit()
-                .putString(SEARCH_HISTORY_KEY, json)
-                .apply()
-            // Toast.makeText(this, "Сохранили значение ${editText.editableText}", Toast.LENGTH_SHORT).show()
+            searchHistory.writeToSharedPreferences()
         }
     }
 
@@ -364,7 +272,7 @@ class SearchActivity : AppCompatActivity() {
         )
         outState.putParcelableArrayList(
             "track_list_search_history",
-            trackListSearchHistory as ArrayList<out Parcelable?>?
+            searchHistory.trackListSearchHistory as ArrayList<out Parcelable?>?
         )
     }
 
@@ -381,10 +289,10 @@ class SearchActivity : AppCompatActivity() {
 
         hidePlaceholder()
 
-        trackListSearchHistory =
+        searchHistory.trackListSearchHistory =
             savedInstanceState.getParcelableArrayList("track_list_search_history")!!
-        trackAdapterSearchHistory.items = trackListSearchHistory
-        trackAdapterSearchHistory.notifyDataSetChanged()
+        searchHistory.trackAdapterSearchHistory.items = searchHistory.trackListSearchHistory
+        searchHistory.trackAdapterSearchHistory.notifyDataSetChanged()
 
         val linearLayoutSearchHistory = findViewById<LinearLayout>(id.searchHistoryLinearLayout)
         linearLayoutSearchHistory.visibility =
@@ -401,5 +309,22 @@ class SearchActivity : AppCompatActivity() {
 
         rvItems.visibility = View.VISIBLE
         placeholderLayout.visibility = View.GONE
+    }
+
+    override fun onItemClick(item: Track) {
+        var itemSearchHistory =
+            searchHistory.trackListSearchHistory.firstOrNull { it.trackId == item.trackId }
+        if (itemSearchHistory != null)
+            searchHistory.trackListSearchHistory.remove(itemSearchHistory)
+        searchHistory.trackListSearchHistory.add(0, item)
+
+        if (searchHistory.trackListSearchHistory.size > 10)
+            searchHistory.trackListSearchHistory.removeAt(10)//(trackListSearchHistory[10])
+
+        searchHistory.trackAdapterSearchHistory.items = searchHistory.trackListSearchHistory
+        searchHistory.trackAdapterSearchHistory.notifyDataSetChanged()
+
+        searchHistory.addItem(item)
+        searchHistory.writeToSharedPreferences()
     }
 }
